% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aid}
\alias{aid}
\title{Analysis of inoculation data}
\usage{
aid(file, type)
}
\arguments{
\item{file}{a text file contains inoculation data.}

\item{type}{grade or lesion, must be specified}
}
\value{
statistical analysis and illustration for inoculation data
}
\description{
Analysis and illustration of inoculation data.
}
\examples{
library(aid)
# grade data
demo1 <- system.file("extdata", "demo1.tsv", package="aid")
aid(demo1, type = "grade")
# lesion data
demo2 <- system.file("extdata", "demo2.tsv", package="aid")
aid(demo2, type = "lesion")
}
\seealso{
\code{\link{grade.test}}, \code{\link{lesion.test}}, \code{\link{grade.barplot}} and \code{\link{lesion.barplot}}.
}

