% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{dsi}
\alias{dsi}
\title{Caculate disease severity}
\usage{
dsi(ino)
}
\arguments{
\item{ino}{a list contains inoculation grade data.}
}
\value{
Disease severity index for individuals.
}
\description{
Caculate disease severity
}
\examples{
demo <- system.file("extdata", "demo1.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
dsi(dat)
}
\seealso{
\code{\link{aid}}, \code{\link{grade.barplot}} and \code{\link{grade.test}}.
}

