% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{error.bar}
\alias{error.bar}
\title{Plot error bar}
\usage{
error.bar(x, y, upper, lower = upper, length = 0.1, ...)
}
\arguments{
\item{x}{position on x-axis.}

\item{y}{position on y-axis.}

\item{upper}{length of upper error bar}
}
\description{
Adding error bar to a barplot
}
\examples{
barx <- barplot(1:10, ylim=c(0,12))
error.bar(barx, 1:10, rep(1,10))
}
\seealso{
\code{\link{aid}}, \code{\link{lesion.barplot}} and \code{\link{biomass.barplot}}.
}

