% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ino.barplot}
\alias{ino.barplot}
\title{Plot inoculation data}
\usage{
ino.barplot(ino)
}
\arguments{
\item{ino}{a list contains inoculation grade data.}
}
\description{
Create a barplot for grade inoculation data
}
\examples{
demo <- system.file("extdata", "demo1.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
ino.barplot(dat)
}
\seealso{
\code{\link{aid}} and \code{\link{ino.colors}}.
}

