% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colors.R
\name{ino.colors}
\alias{ino.colors}
\title{The inoculation colors vector}
\usage{
ino.colors(n, alpha = 1)
}
\arguments{
\item{n}{number of colors, usually equal to (or greater that) the number of desease grades.}

\item{alpha}{the alpha transparency, a number in [0,1]}
}
\value{
a vector of contiguous colors.
}
\description{
Create a color vector.
}
\examples{
ino.colors(10)
plot(1:5, pch = 16, col = ino.colors(5), cex=3)
}
\seealso{
\code{\link{aid}} and \code{\link{grade.barplot}}.
}

