% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lesion.barplot}
\alias{lesion.barplot}
\title{Plot lesion data}
\usage{
lesion.barplot(ino)
}
\arguments{
\item{ino}{a list contains inoculation lesion data.}
}
\description{
Create a barplot for lesion inoculation data
}
\examples{
demo <- system.file("extdata", "demo2.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
lesion.barplot(dat)
}
\seealso{
\code{\link{aid}}, \code{\link{lesion.test}}, \code{\link{biomass.barplot}} and \code{\link{grade.barplot}}.
}

