% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.ino}
\alias{plot.ino}
\title{Plot inoculation data}
\usage{
plot.ino(ino)
}
\arguments{
\item{a}{list contains inoculation grade data.}
}
\description{
Create a barplot for grade inoculation data
}
\examples{
dat <- read.table(file, header = TRUE, check.names=FALSE)
plot.ino(dat)
}
\seealso{
\code{\link{aid}} and \code{\link{ino.colors}}.
}

