% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sym.pval}
\alias{sym.pval}
\title{Assign p-value symbols}
\usage{
sym.pval(pval)
}
\arguments{
\item{pval}{pvalue, between 0-1.}
}
\description{
Convert p-value to symbols
}
\examples{
pval <- c(0.15, 0.10, 0.05, 0.02, 0.01, 0.0001)
n <- length(pval)
sym  <- rep(" ", n)
for (i in 1:n)
  sym[i] <- sym.pval(pval[i])
cat(sym)
}
\seealso{
\code{\link{aid}}, \code{\link{grade.test}} and \code{\link{lesion.test}}.
}

