\name{AID-package}
\alias{AID-package}
\alias{AID-package}
\docType{package}
\title{
Box-Cox Power Transformation
}
\description{
Performs Box-Cox power transformation for different purposes, graphical approaches, assess the success of the transformation via tests and plots, computes mean and confidence interval for back transformed data.
}
\details{
\tabular{ll}{
Package: \tab AID\cr
Type: \tab Package\cr
License: \tab GPL (>=2)\cr
}
}