\name{AID}
\alias{AID}
\alias{AID}
\docType{package}
\title{
An R Package to Estimate Box-Cox Power Transformation Parameter 
}
\description{
Includes a function to estimate the power transformation parameter and some datasets
}
\details{
\tabular{ll}{
Package: \tab AID\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2013-04-29\cr
License: \tab GPL (>=2)\cr
}
}