\name{AID-package}
\alias{AID-package}
\alias{AID-package}
\docType{package}
\title{
Estimation of Box-Cox Power Transformation Parameter 
}
\description{
Power transformation parameter is estimated for different purposes.
}
\details{
\tabular{ll}{
Package: \tab AID\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2015-04-09\cr
License: \tab GPL (>=2)\cr
}
}