\name{AID-package}
\alias{AID-package}
\alias{AID-package}
\docType{package}
\title{
Box-Cox Power Transformation
}
\description{
Box-Cox power transformation parameter is estimated for different purposes.
}
\details{
\tabular{ll}{
Package: \tab AID\cr
Type: \tab Package\cr
License: \tab GPL (>=2)\cr
}
}