% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aid}
\alias{aid}
\title{Analysis of inoculation data}
\usage{
aid(file, type)
}
\arguments{
\item{file}{a text file contains inoculation data.}
}
\value{
statistical analysis and illustration for inoculation data
}
\description{
Analysis and illustration of inoculation data.
}
\examples{
library(aid)
demo <- system.file("extdata", "demo1.tsv", package="aid")
aid(demo)
}
\seealso{
\code{\link{grade.test}} and \code{\link{grade.barplot}}.
}

