% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aid}
\alias{aid}
\title{Analysis of inoculation data}
\usage{
aid(file)
}
\arguments{
\item{a}{text file contain inoculation data.}
}
\value{
statistical analysis and illustration for inoculation data
}
\description{
Analysis and illustration of inoculation data.
}
\examples{
library(aid)
aid("~/ino.data")
}
\seealso{
\code{\link{grade.test}} and \code{\link{plot.ino}}.
}

