% Generated by roxygen2 (4.0.2): do not edit by hand
\name{biomass.barplot}
\alias{biomass.barplot}
\title{Plot biomass data}
\usage{
biomass.barplot(ino)
}
\arguments{
\item{ino}{a list contains inoculation biomass data.}
}
\description{
Create a barplot for biomass (qPCR) inoculation data
}
\examples{
demo <- system.file("extdata", "demo3.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
biomass.barplot(dat)
}
\seealso{
\code{\link{aid}}, \code{\link{biomass.test}}, \code{\link{lesion.barplot}} and \code{\link{grade.barplot}}.
}

