% Generated by roxygen2 (4.0.2): do not edit by hand
\name{biomass.test}
\alias{biomass.test}
\title{Statistical test for biomass data}
\usage{
biomass.test(bio)
}
\arguments{
\item{ino}{a list contains inoculation biomass data.}
}
\value{
p-value symbols indicating the significances between comparisons.
}
\description{
Perform t test on biomass (qPCR) data.
}
\examples{
bio<-c(0.82, 3.14, 0.88, 3.21, 0.85, 3.20)
dim(bio)<-c(2,3)
biomass.test(bio)
}
\seealso{
\code{\link{aid}}, \code{\link{biomass.barplot}}, \code{\link{lesion.test}} and \code{\link{grade.test}}.
}

