% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grade.barplot}
\alias{grade.barplot}
\title{Plot inoculation data}
\usage{
grade.barplot(ino)
}
\arguments{
\item{ino}{a list contains inoculation grade data.}
}
\description{

}
\details{
Create a barplot for grade inoculation data
}
\examples{
demo <- system.file("extdata", "demo1.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
grade.barplot(dat)
}
\seealso{
\code{\link{aid}} and \code{\link{ino.colors}}.
}

