% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grade.barplot}
\alias{grade.barplot}
\title{Plot grade data}
\usage{
grade.barplot(ino)
}
\arguments{
\item{ino}{a list contains inoculation grade data.}
}
\description{
Create a barplot for grade inoculation data
}
\examples{
demo <- system.file("extdata", "demo1.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
grade.barplot(dat)
}
\seealso{
\code{\link{aid}}, \code{\link{ino.colors}}, \code{\link{grade.test}}, \code{\link{lesion.barplot}} and \code{\link{biomass.barplot}}..
}

