% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grade.test}
\alias{grade.test}
\title{Statistical test for grade data}
\usage{
grade.test(ino)
}
\arguments{
\item{ino}{a list contains inoculation grade data.}
}
\value{
p-value signs indicating the significances between comparisons.
}
\description{
Convert grade data to real observations and perform t test.
}
\examples{
demo <- system.file("extdata", "demo1.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
grade.test(dat)
}
\seealso{
\code{\link{aid}}, and \code{\link{grade.barplot}}.
}

