% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.R
\name{grade.test}
\alias{grade.test}
\title{Statistical test for grade data}
\usage{
grade.test(ino, alternative)
}
\arguments{
\item{ino}{a list contains inoculation grade data.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
p-value symbols indicating the significances between comparisons.
}
\description{
Convert grade data to real observations and perform t test.
}
\examples{
demo <- system.file("extdata", "demo1.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
grade.test(dat, alternative = "t")
}
\seealso{
\code{\link{aid}}, \code{\link{grade.barplot}}, \code{\link{lesion.test}} and \code{\link{biomass.test}}.
}

