% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lesion.barplot}
\alias{lesion.barplot}
\title{Plot lesion data}
\usage{
lesion.barplot(ino, alternative)
}
\arguments{
\item{ino}{a list contains inoculation lesion data.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\description{
Create a barplot for lesion inoculation data
}
\examples{
demo <- system.file("extdata", "demo2.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
lesion.barplot(dat, alternative)
}
\seealso{
\code{\link{aid}}, \code{\link{lesion.test}}, \code{\link{biomass.barplot}} and \code{\link{grade.barplot}}.
}

