% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.R
\name{lesion.test}
\alias{lesion.test}
\title{Statistical test for lesion data}
\usage{
lesion.test(ino, alternative)
}
\arguments{
\item{ino}{a list contains inoculation lesion data.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
p-value symbols indicating the significances between comparisons.
}
\description{
Perform t test on lesion data.
}
\examples{
demo <- system.file("extdata", "demo2.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
lesion.test(dat, alternative = "t")
}
\seealso{
\code{\link{aid}}, \code{\link{lesion.barplot}}, \code{\link{biomass.test}} and \code{\link{grade.test}}.
}

