% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lesion.test}
\alias{lesion.test}
\title{Statistical test for lesion data}
\usage{
lesion.test(ino)
}
\arguments{
\item{ino}{a list contains inoculation lesion data.}
}
\value{
p-value symbols indicating the significances between comparisons.
}
\description{
Perform t test on lesion data.
}
\examples{
demo <- system.file("extdata", "demo2.tsv", package="aid")
dat <- read.table(demo, header = TRUE, check.names=FALSE)
lesion.test(dat)
}
\seealso{
\code{\link{aid}}, \code{\link{grade.test}} and \code{\link{lesion.barplot}}.
}

